Added new api route for previous authentication implementation. Also renamed html page I was using for that auth method.
