+++ routes/htmlRoutes/index.js
@@ -1,7 +1,10 @@
 const router = require(`express`).Router();
+const path = require("path");
 
 router.get(`/`, function (req, res) {
-  res.status(200).sendFile(`../../public/index.html`);
+  res
+    .status(200)
+    .sendFile(path.join(__dirname, `../../public/createAsanaTask.html`));
 });
 
 module.exports = router;
